Changes to required permissions for project/service add/edit
Uses the @user_passes_test lambda to check that users have the required 'rse' permissions.
Prior to this the requirement was simply being logged in.
This allows per RSE permissions to allow/deny add/edit of allocated projects and/or service projects.

The project_new() method handles both allocated projects and service projects.
The initial entry page (with drop down to choose between types) is viewable if the RSE has one of the permissions to add.
Subsequent project_new_allocated() and project_new_service are then tested individually.

The project_edit() method does not defer to further methods for each type.
Instead the code checks the instance type and differs in response.
The appropriate permssion tests are added to the isinstance branches and a redirect to 'login' occurs if permission is not granted.